Move result objects to another file; eliminate needless state update
